requirement.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
algoliasearch==2.6.3
algoliasearch-django==2.0.0
altgraph==0.17.3
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
auto-py-to-exe==2.27.0
backports.zoneinfo==0.2.1
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.7
crispy-bootstrap4==2022.1
cryptography==39.0.1
defusedxml==0.7.1
distro==1.8.0
Django==4.1.5
django-allauth==0.54.0
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-extensions==3.2.3
django-filter==22.1
django-phonenumber-field==7.1.0
django-twilio==0.14.3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
EasyNMT==2.0.2
Eel==0.14.0
exceptiongroup==1.2.0
fasttext==0.9.2
filelock==3.13.1
frozenlist==1.4.0
fsspec==2024.2.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
greenlet==2.0.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.21.1
idna==3.4
Jinja2==3.1.3
joblib==1.3.2
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.0.4
mysqlclient==2.2.0
networkx==3.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai==1.3.5
packaging==23.2
phonenumbers==8.13.18
Pillow==10.0.0
protobuf==4.25.3
pybind11==2.11.1
pycparser==2.21
pydantic==2.5.2
pydantic-core==2.14.5
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
PyJWT==2.6.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
regex==2023.12.25
requests==2.28.2
requests-oauthlib==1.3.1
rest-framework-pagination==0.0.1
sacremoses==0.1.1
safetensors==0.4.2
sentencepiece==0.2.0
sniffio==1.3.0
sqlparse==0.4.3
sympy==1.12
tokenizers==0.15.2
torch==2.2.1
tqdm==4.66.1
transformers==4.38.1
triton==2.2.0
twilio==7.17.0
typing-extensions==4.10.0
urllib3==1.26.14
whichcraft==0.6.1
yarl==1.9.2
zope.event==4.6
zope.interface==5.5.2
